+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.Rss_aggregator2022"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.moaimar.rss_aggregator_2022.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
