--- app/src/main/java/com/example/rss_aggregator_2022/app/domain/ErrorApp.kt
@@ -1,6 +1,0 @@
-package com.example.rss_aggregator_2022.app.domain
-
-sealed class ErrorApp {
-    class DataError() : ErrorApp()
-    class DomainError() : ErrorApp()
-}
