--- app/src/main/java/com/example/rss_aggregator_2022/features/rssfeed/presentation/RssFeedFragment.kt
@@ -1,21 +1,0 @@
-package com.example.rss_aggregator_2022.features.rssfeed.presentation
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.rss_aggregator_2022.databinding.FragmentRssFeedBinding
-
-class RssFeedFragment : Fragment() {
-    var binding: FragmentRssFeedBinding? = null
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentRssFeedBinding.inflate(inflater)
-        return binding?.root
-    }
-}
