--- app/src/main/java/com/example/rss_aggregator_2022/features/rssmanagement/data/RssDataRepository.kt
@@ -1,20 +1,0 @@
-package com.example.rss_aggregator_2022.features.rssmanagement.data
-
-import com.example.rss_aggregator_2022.app.domain.ErrorApp
-import com.example.rss_aggregator_2022.app.functional.Either
-import com.example.rss_aggregator_2022.features.rssmanagement.data.local.LocalDataSource
-import com.example.rss_aggregator_2022.features.rssmanagement.domain.Rss
-import com.example.rss_aggregator_2022.features.rssmanagement.domain.RssRepository
-import kotlinx.coroutines.flow.Flow
-
-class RssDataRepository(private val localSource: LocalDataSource) : RssRepository {
-    override suspend fun createRss(urlRss: String, name: String) =
-        localSource.saveRss(urlRss, name)
-
-    override suspend fun getSourceRss(): Flow<Either<ErrorApp, List<Rss>>> =
-        localSource.obtain()
-
-    override suspend fun deleteRss(urlRss: String) {
-        localSource.delete(urlRss)
-    }
-}
