--- app/src/main/java/com/example/rss_aggregator_2022/features/rssmanagement/domain/Model.kt
@@ -1,3 +1,0 @@
-package com.example.rss_aggregator_2022.features.rssmanagement.domain
-
-data class Rss(val urlRss: String, val name: String)
