--- app/src/main/java/com/example/rss_aggregator_2022/features/rssmanagement/domain/SaveRssUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.rss_aggregator_2022.features.rssmanagement.domain
-
-class SaveRssUseCase(private val repository: RssRepository) {
-    suspend operator fun invoke(urlRss: String, name: String) {
-        repository.createRss(urlRss, name)
-    }
-}
