+++ app/src/main/java/com/moaimar/rss_aggregator_2022/app/domain/ErrorApp.kt
@@ -1,0 +1,6 @@
+package com.moaimar.rss_aggregator_2022.app.domain
+
+sealed class ErrorApp {
+    class DataError() : ErrorApp()
+    class DomainError() : ErrorApp()
+}
