+++ app/src/main/java/com/moaimar/rss_aggregator_2022/features/profile/ProfileFragment.kt
@@ -1,0 +1,21 @@
+package com.moaimar.rss_aggregator_2022.features.profile
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.rss_aggregator_2022.databinding.FragmentProfileBinding
+
+class ProfileFragment : Fragment() {
+    var binding: FragmentProfileBinding? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentProfileBinding.inflate(inflater)
+        return binding?.root
+    }
+}
