+++ app/src/main/java/com/moaimar/rss_aggregator_2022/features/rssmanagement/domain/DeleteSourceRssUseCase.kt
@@ -1,0 +1,7 @@
+package com.moaimar.rss_aggregator_2022.features.rssmanagement.domain
+
+class DeleteSourceRssUseCase(private val repository: RssRepository) {
+    suspend operator fun invoke(urlRss: String) {
+        repository.deleteRss(urlRss)
+    }
+}
