+++ app/src/main/java/com/moaimar/rss_aggregator_2022/features/rssmanagement/domain/Model.kt
@@ -1,0 +1,3 @@
+package com.moaimar.rss_aggregator_2022.features.rssmanagement.domain
+
+data class Rss(val urlRss: String, val name: String)
